refactor(suite-native): Styles
