+++ suite-native/module-home/src/components/DashboardHeader.tsx
@@ -1,23 +1,29 @@
 import React from 'react';
 
+import { prepareNativeStyle, useNativeStyles } from '@trezor/styles';
 import { Icon } from '@trezor/icons';
 import { Box } from '@suite-native/atoms';
 
 import { DashboardHeaderDeviceChip } from './DashboardHeaderDeviceChip';
 
-export const DashboardHeader = () => (
-    <Box
-        flexDirection="row"
-        justifyContent="space-between"
-        alignItems="center"
-        style={{ width: '100%' }}
-    >
-        <DashboardHeaderDeviceChip />
-        <Box flexDirection="row">
-            <Box marginRight="medium">
-                <Icon name="eyeSlash" />
+const headerStyle = prepareNativeStyle(() => ({
+    width: '100%',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+}));
+
+export const DashboardHeader = () => {
+    const { applyStyle } = useNativeStyles();
+    return (
+        <Box style={applyStyle(headerStyle)}>
+            <DashboardHeaderDeviceChip />
+            <Box flexDirection="row">
+                <Box marginRight="medium">
+                    <Icon name="eyeSlash" />
+                </Box>
+                <Icon name="notifications" />
             </Box>
-            <Icon name="notifications" />
         </Box>
-    </Box>
-);
+    );
+};
